Fixed form2js version number.
FR used a special version of form2js called '2.0-769718a'. The last part
seems to be a commit hash: maxatwork/form2js@769718a . Version 2.0 is
actually older then the version FR used. It is assumed they bumped to
this version because they needed a fix or improvement which wasn't
available in a "properly" released version.